build/feat: Linux compilation fixes — pointer casts, OpenAL audio, SDL3 input

+++ Generals/Code/GameEngine/Source/GameLogic/AI/AIStates.cpp
@@ -1229,7 +1229,8 @@ Bool outOfWeaponRangePosition( State *thisState, void* userData )
  */
 static Bool cannotPossiblyAttackObject( State *thisState, void* userData )
 {
-	AbleToAttackType attackType = (AbleToAttackType)(UnsignedInt)userData;
+	// GeneralsX @build Fix pointer-to-integer truncation on 64-bit platforms.
+	AbleToAttackType attackType = (AbleToAttackType)(UnsignedInt)(uintptr_t)userData;
 	Object *obj = thisState->getMachineOwner();
 	Object *victim = thisState->getMachineGoalObject();
